fix: alert 에러 해결
- 동시에 alert 두 개 선언 불가
- 하나의 modifier에서 분기 처리를 한다.

+++ Dal_Da/Dal_Da/HomeView/CarouselCardView/MoonCardView.swift
@@ -13,8 +13,11 @@ struct MoonCardView: View {
     @Environment(\.modelContext) private var modelContext
     
     var moon: Moon
+    
     @State private var showingDeleteAlert = false
     @State private var showingSaveAlert = false
+    
+    @State private var activeAlert = false
     
     var body: some View {
         ZStack(alignment: .bottomLeading) {
@@ -30,7 +33,8 @@ struct MoonCardView: View {
                         Menu {
                             
                             Button {
-                                showingSaveAlert.toggle()
+                                activeAlert.toggle()
+                                self.showingSaveAlert.toggle()
                                 print("Save~~~~~~~~~~~")
                             } label: {
                                 Label("Save", systemImage: "square.and.arrow.down")
@@ -38,10 +42,12 @@ struct MoonCardView: View {
                             
                             
                             
-                            Divider()
+                            //                            Divider()
                             
                             Button(role: .destructive) {
-                                showingDeleteAlert.toggle()
+                                print("누ㅠㄹ림ㄴㅇㄹㅁ닝러ㅣ마넝ㄹ")
+                                activeAlert.toggle()
+                                self.showingDeleteAlert.toggle()
                                 
                             } label: {
                                 Label("Delete", systemImage: "trash")
@@ -51,21 +57,33 @@ struct MoonCardView: View {
                             Image("Three Dots")
                         }
                         .padding()
-                        .alert(isPresented: $showingDeleteAlert) {
-                            Alert(title: Text("이미지 삭제"), message: Text("삭제된 이미지는 복구할 수 없습니다.\n정말로 삭제하시겠습니까?"), primaryButton: .destructive(Text("삭제"), action: {
-                                //some Action
-                                modelContext.delete(moon)
-                            }), secondaryButton: .cancel(Text("취소")))
-                        }
-                        .alert(isPresented: $showingSaveAlert) {
-                            Alert(title: Text("이미지 저장"), message: Text("이 이미지를 사진 앨범에 저장하려고 합니다.\n저장하시겠습니까?"), primaryButton: .default(Text("저장"), action: {
-                                //some Action
-                                savePhoto(uiImage)
-                            }), secondaryButton: .cancel(Text("취소")))
-                        }
                         
-                        
+                        .alert(isPresented: $activeAlert) {
+                            
+                            if showingDeleteAlert {
+//                                showingDeleteAlert.toggle()
+                                
+                                Alert(title: Text("이미지 삭제"), message: Text("삭제된 이미지는 복구할 수 없습니다.\n정말로 삭제하시겠습니까?"), primaryButton: .destructive(Text("삭제"), action: {
+                                    //some Action
+                                    modelContext.delete(moon)
+                                }), secondaryButton: .cancel(Text("취소"), action: {
+                                    showingDeleteAlert.toggle()
+                                }))
+                                
+                            } else {
+//                                showingSaveAlert.toggle()
+                                Alert(title: Text("이미지 저장"), message: Text("이 이미지를 사진 앨범에 저장하려고 합니다.\n저장하시겠습니까?"), primaryButton: .default(Text("저장"), action: {
+                                    //some Action
+                                    savePhoto(uiImage)
+                                }), secondaryButton: .cancel(Text("취소"), action: {
+                                    showingSaveAlert.toggle()
+                                }))
+                                
+                            }
+                            
+                        }
                     }
+                        
                     .cornerRadius(24)
                 
             } else {
@@ -91,15 +109,15 @@ struct MoonCardView: View {
                         .foregroundColor(.white)
                     
                     // 이미지 하단에 위치할 텍스트
-                    VStack(alignment: .leading,spacing:4) {
+                    VStack(alignment: .leading, spacing: 6) {
                         
                         Text("\(DateUtilities.formatDateTime(date, formatType: "EEEE"))\(moon.shape == "" ? "" : ", \(moon.shape)")")
-                            .font(.system(size: 16, weight: .regular))
+                            .font(.system(size: 13, weight: .regular))
                             .foregroundColor(.white)
                         
                         Text(moon.memo)
                             .font(.system(size: 16, weight: .regular))
-                            .foregroundColor(.white)
+                            .foregroundColor(.gray100)
                         
                     }
